+++ CHANGELOG.md
@@ -3,7 +3,13 @@
 ### New Features : 
 
 #### HLSL Library
-* Bezier Library. ([]())
+* Bezier Library. ([02d2bb9](https://github.com/NazioLT/NZCore/commit/02d2bb97e581ca29d9a4d987394d3a32e747debd))
+
+#### Other : 
+* Libraries Cleaning.
+
+#### Removings : 
+* Instantiator Library.
 
 ## [1.2.1](https://github.com/NazioLT/NZCore/releases/tag/1.2.1) (2023-01-24)
 
@@ -32,7 +38,7 @@
 
 ## [1.1.0](https://github.com/NazioLT/NZCore/releases/tag/1.1.0) (2022-12-8)
 
-### Bug solving :
+### Issue solving :
 
 * NCurveBehaviour scene blocking. ([ea414e6](https://github.com/NazioLT/NZCore/commit/ea414e640570124c5aa807dcfd6d13a7986a7557))
 * NCurveBehaviour mesh generation errors. ([ea414e6](https://github.com/NazioLT/NZCore/commit/ea414e640570124c5aa807dcfd6d13a7986a7557))